Move container class to the base template
This was redundant on the page templates and this way its easier to add some spacer helper classes for padding top and padding bottom. The top just makes it look nicer, the bottom padding helps with spacing on mobile (which is also solved by removing the stickcy elements in a later commit).