Move github link from sticky to navbar, split navbar
The sticky github link overlapped some interaction elements on mobile (tested with iPhone, Chrome dev tools). The easiest way to solve this seems to be to just move the link to the navigation, since there is no concept of a footer, all links are at the top ATM. So why not the github Link as well.

To make it easier to parse, I split the navbar. Left are feautre links, right are kind of service links.